Combine rely and deny transformers into auth transformer
