+++ environments/docker.toml
@@ -180,7 +180,7 @@
         contracts = ["MCD_FLIP_ETH_A", "MCD_FLIP_BAT_A", "MCD_FLIP_SAI", "MCD_FLOP"]
         rank = "0"
     [exporter.deny]
-        path = "transformers/events/deny/initializer"
+        path = "transformers/events/auth/deny_initializer"
         type = "eth_event"
         repository = "github.com/makerdao/vdb-mcd-transformers"
         migrations = "db/migrations"
@@ -300,7 +300,7 @@
         contracts = ["MCD_POT"]
         rank = "0"
     [exporter.rely]
-        path = "transformers/events/rely/initializer"
+        path = "transformers/events/auth/rely_initializer"
         type = "eth_event"
         repository = "github.com/makerdao/vdb-mcd-transformers"
         migrations = "db/migrations"
@@ -350,7 +350,7 @@
         contracts = ["MCD_FLIP_ETH_A", "MCD_FLIP_BAT_A", "MCD_FLIP_SAI", "MCD_FLAP", "MCD_FLOP"]
         rank = "0"
     [exporter.vat_deny]
-        path = "transformers/events/deny/vat_initializer"
+        path = "transformers/events/auth/vat_deny_initializer"
         type = "eth_event"
         repository = "github.com/makerdao/vdb-mcd-transformers"
         migrations = "db/migrations"
@@ -427,7 +427,7 @@
         contracts = ["MCD_VAT"]
         rank = "0"
     [exporter.vat_rely]
-        path = "transformers/events/rely/vat_initializer"
+        path = "transformers/events/auth/vat_rely_initializer"
         type = "eth_event"
         repository = "github.com/makerdao/vdb-mcd-transformers"
         migrations = "db/migrations"
