+++ transformers/events/auth/auth_suite_test.go
@@ -1,4 +1,4 @@
-package rely_test
+package auth_test
 
 import (
 	"testing"
@@ -7,7 +7,7 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-func TestRely(t *testing.T) {
+func TestAuth(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "Rely Suite")
+	RunSpecs(t, "Auth Suite")
 }
