+++ transformers/events/auth/deny_initializer/initializer.go
@@ -1,7 +1,7 @@
-package vat_initializer
+package deny_initializer
 
 import (
-	"github.com/makerdao/vdb-mcd-transformers/transformers/events/deny"
+	"github.com/makerdao/vdb-mcd-transformers/transformers/events/auth"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/shared"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/shared/constants"
 	"github.com/makerdao/vulcanizedb/libraries/shared/factories/event"
@@ -10,5 +10,5 @@ import (
 
 var EventTransformerInitializer transformer.EventTransformerInitializer = event.ConfiguredTransformer{
 	Config:      shared.GetEventTransformerConfig(constants.DenyTable, constants.DenySignature()),
-	Transformer: deny.Transformer{LogNoteArgumentOffset: -1},
+	Transformer: auth.Transformer{TableName: constants.DenyTable},
 }.NewTransformer
