+++ transformers/events/auth/vat_deny_initializer/initializer.go
@@ -1,0 +1,14 @@
+package vat_deny_initializer
+
+import (
+	"github.com/makerdao/vdb-mcd-transformers/transformers/events/auth"
+	"github.com/makerdao/vdb-mcd-transformers/transformers/shared"
+	"github.com/makerdao/vdb-mcd-transformers/transformers/shared/constants"
+	"github.com/makerdao/vulcanizedb/libraries/shared/factories/event"
+	"github.com/makerdao/vulcanizedb/libraries/shared/transformer"
+)
+
+var EventTransformerInitializer transformer.EventTransformerInitializer = event.ConfiguredTransformer{
+	Config:      shared.GetEventTransformerConfig(constants.DenyTable, constants.DenySignature()),
+	Transformer: auth.Transformer{TableName: constants.DenyTable, LogNoteArgumentOffset: -1},
+}.NewTransformer
