+++ transformers/events/auth/vat_rely_initializer/initializer.go
@@ -1,7 +1,7 @@
-package initializer
+package vat_rely_initializer
 
 import (
-	"github.com/makerdao/vdb-mcd-transformers/transformers/events/deny"
+	"github.com/makerdao/vdb-mcd-transformers/transformers/events/auth"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/shared"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/shared/constants"
 	"github.com/makerdao/vulcanizedb/libraries/shared/factories/event"
@@ -9,6 +9,6 @@ import (
 )
 
 var EventTransformerInitializer transformer.EventTransformerInitializer = event.ConfiguredTransformer{
-	Config:      shared.GetEventTransformerConfig(constants.DenyTable, constants.DenySignature()),
-	Transformer: deny.Transformer{},
+	Config:      shared.GetEventTransformerConfig(constants.RelyTable, constants.RelySignature()),
+	Transformer: auth.Transformer{TableName: constants.RelyTable, LogNoteArgumentOffset: -1},
 }.NewTransformer
