--- transformers/events/deny/deny_suite_test.go
@@ -1,13 +1,0 @@
-package deny_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestDeny(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Deny Suite")
-}
