--- transformers/events/deny/transformer_test.go
@@ -1,77 +1,0 @@
-package deny_test
-
-import (
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/makerdao/vdb-mcd-transformers/test_config"
-	"github.com/makerdao/vdb-mcd-transformers/transformers/events/deny"
-	"github.com/makerdao/vdb-mcd-transformers/transformers/shared"
-	"github.com/makerdao/vdb-mcd-transformers/transformers/shared/constants"
-	"github.com/makerdao/vdb-mcd-transformers/transformers/test_data"
-	"github.com/makerdao/vulcanizedb/libraries/shared/factories/event"
-	"github.com/makerdao/vulcanizedb/pkg/core"
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-var _ = Describe("Deny Transformer", func() {
-	var db = test_config.NewTestDB(test_config.NewTestNode())
-
-	BeforeEach(func() {
-		test_config.CleanTestDB(db)
-	})
-
-	It("converts normal deny logs to models", func() {
-		converter := deny.Transformer{}
-		models, err := converter.ToModels(constants.CatABI(), []core.HeaderSyncLog{test_data.DenyHeaderSyncLog}, db)
-		Expect(err).NotTo(HaveOccurred())
-
-		var contractAddressID int64
-		contractAddressErr := db.Get(&contractAddressID, `SELECT id FROM addresses WHERE address = $1`,
-			test_data.DenyHeaderSyncLog.Log.Address.String())
-		Expect(contractAddressErr).NotTo(HaveOccurred())
-
-		var usrAddressID int64
-		usrAddressErr := db.Get(&usrAddressID, `SELECT id FROM addresses WHERE address = $1`,
-			common.HexToAddress(test_data.DenyHeaderSyncLog.Log.Topics[2].Hex()).Hex())
-		Expect(usrAddressErr).NotTo(HaveOccurred())
-
-		expectedModel := test_data.DenyModel()
-		expectedModel.ColumnValues[event.AddressFK] = contractAddressID
-		expectedModel.ColumnValues[constants.UsrColumn] = usrAddressID
-
-		Expect(models).To(Equal([]event.InsertionModel{expectedModel}))
-	})
-
-	It("converts Vat deny logs to models", func() {
-		converter := deny.Transformer{LogNoteArgumentOffset: -1}
-		models, err := converter.ToModels(constants.VatABI(), []core.HeaderSyncLog{test_data.VatDenyHeaderSyncLog}, db)
-		Expect(err).NotTo(HaveOccurred())
-
-		var contractAddressID int64
-		contractAddressErr := db.Get(&contractAddressID, `SELECT id FROM addresses WHERE address = $1`,
-			test_data.VatDenyHeaderSyncLog.Log.Address.String())
-		Expect(contractAddressErr).NotTo(HaveOccurred())
-
-		var usrAddressID int64
-		usrAddressErr := db.Get(&usrAddressID, `SELECT id FROM addresses WHERE address = $1`,
-			common.HexToAddress(test_data.VatDenyHeaderSyncLog.Log.Topics[1].Hex()).Hex())
-		Expect(usrAddressErr).NotTo(HaveOccurred())
-
-		expectedModel := test_data.VatDenyModel()
-		expectedModel.ColumnValues[event.AddressFK] = contractAddressID
-		expectedModel.ColumnValues[constants.UsrColumn] = usrAddressID
-
-		Expect(models).To(Equal([]event.InsertionModel{expectedModel}))
-	})
-
-	It("returns an error if the expected amount of topics aren't in the log", func() {
-		converter := deny.Transformer{}
-		invalidLog := test_data.DenyHeaderSyncLog
-		invalidLog.Log.Topics = []common.Hash{}
-
-		_, err := converter.ToModels(constants.CatABI(), []core.HeaderSyncLog{invalidLog}, db)
-
-		Expect(err).To(HaveOccurred())
-		Expect(err).To(MatchError(shared.ErrLogMissingTopics(3, 0)))
-	})
-})
